+++ flutter_badge_manager_android/example/lib/local_notifications_manager.dart
@@ -1,5 +1,5 @@
 // autor - <a.a.ustinoff@gmail.com> Anton Ustinoff
-// ignore_for_file: sort_constructors_first
+// ignore_for_file: public_member_api_docs, sort_constructors_first
 
 import 'dart:async';
 import 'dart:io' as io;
@@ -13,9 +13,9 @@ void _onTapBackgroundNotification(NotificationResponse? notificationResponse) {}
 /// LocalNotificationsManager class
 /// {@endtemplate}
 final class LocalNotificationsManager {
-  static final _internalSingleton = LocalNotificationsManager._internal();
-  static LocalNotificationsManager get instance => _internalSingleton;
-  factory LocalNotificationsManager() => _internalSingleton;
+  static final _instance = LocalNotificationsManager._internal();
+  static LocalNotificationsManager get instance => _instance;
+  factory LocalNotificationsManager() => _instance;
   LocalNotificationsManager._internal() {
     _flutterLocalNotificationsPlugin = FlutterLocalNotificationsPlugin();
     _setupLocalNotificationPlugin();
@@ -33,9 +33,6 @@ final class LocalNotificationsManager {
   late final FlutterLocalNotificationsPlugin _flutterLocalNotificationsPlugin;
   // FirebaseMessaging get messaging => FirebaseMessaging.instance;
 
-  // final StreamController<ReceivedNotification?> _receivedNotificationsStreamController = StreamController<ReceivedNotification?>.broadcast();
-  // Stream<ReceivedNotification?> get didReceivedNotifications => _receivedNotificationsStreamController.stream;
-
   final StreamController<NotificationResponse?> _notificationsStreamController =
       StreamController<NotificationResponse?>.broadcast();
 
@@ -46,7 +43,6 @@ final class LocalNotificationsManager {
   /// Dispose all streams
   void dispose() {
     _notificationsStreamController.close();
-    // _receivedNotificationsStreamController.close();
   }
 
   /// Show local notification
@@ -65,41 +61,6 @@ final class LocalNotificationsManager {
       payload: payload,
     );
   }
-
-  /// Update badge count on Android
-  /// Used only to silently update badge count
-  /// Used because of the lack of a way to update the badge count on Android
-  /* Future<void> updateBadgeCount$Android(int count) async {
-    final androidDetails = AndroidNotificationDetails(
-      'badge_channel_id',
-      'Badge Channel',
-      channelDescription: 'Used only to silently update badge count',
-      importance: Importance.defaultImportance,
-      priority: Priority.defaultPriority,
-      showWhen: false,
-      playSound: false,
-      enableVibration: false,
-      channelShowBadge: true,
-      number: count,
-    );
-
-    final notificationDetails = NotificationDetails(
-      android: androidDetails,
-      iOS: DarwinNotificationDetails(
-        presentAlert: false,
-        presentBadge: true,
-        badgeNumber: count,
-        presentSound: false,
-      ),
-    );
-
-    await _flutterLocalNotificationsPlugin.show(
-      9999,
-      ' ',
-      null,
-      notificationDetails,
-    );
-  } */
 
   /// Get [NotificationDetails]
   Future<NotificationDetails?> _getNotificationDetails() async {
@@ -140,19 +101,7 @@ final class LocalNotificationsManager {
     );
     await _flutterLocalNotificationsPlugin.initialize(
       initializationSettings,
-      onDidReceiveNotificationResponse: (details) {
-        _notificationsStreamController.add(details);
-        /* switch (notificationResponse.notificationResponseType) {
-          case NotificationResponseType.selectedNotification:
-            _notificationsStreamController.add(notificationResponse.payload);
-            break;
-          case NotificationResponseType.selectedNotificationAction:
-            if (notificationResponse.actionId == navigationActionId) {
-              _notificationsStreamController.add(notificationResponse.payload);
-            }
-            break;
-        } */
-      },
+      onDidReceiveNotificationResponse: _notificationsStreamController.add,
       onDidReceiveBackgroundNotificationResponse: _onTapBackgroundNotification,
     );
     // await FirebaseMessaging.instance